Restrict modifying courses to authorized users

+++ courses/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, reverse, redirect, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 from .models import Course, CourseTier
 from .forms import CourseForm
@@ -38,8 +39,13 @@ def course_detail(request, course_id):
     return render(request, 'courses/course_detail.html', context)
 
 
+@login_required
 def add_course(request):
     """ Add a course to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
+        
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES)
         course_name = request.POST.get('name', 'course')
@@ -60,8 +66,13 @@ def add_course(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_course(request, course_id):
     """ Edit a course in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
+
     course = get_object_or_404(Course, pk=course_id)
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES, instance=course)
@@ -84,8 +95,13 @@ def edit_course(request, course_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_course(request, course_id):
     """ Delete a course from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admin can do that.')
+        return redirect(reverse('home'))
+
     course = get_object_or_404(Course, pk=course_id)
     course_name = course.name 
     course.delete()
